requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
chromedriver-installer==0.0.6
cryptography==2.2.2
idna==2.6
ndg-httpsclient==0.4.4
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
PyYAML==3.12
requests==2.18.4
selenium==3.11.0
six==1.11.0
urllib3==1.22
